create circuits_utils to simplify further
